fix: working GPU compute pipeline (FFI by-ptr wrappers, ASAP destruction pins, enum values)
Critical fixes for real GPU execution:

1. FFI calling convention: Mojo's DLHandle.call cannot pass >16-byte
   structs by value (x86_64 SysV ABI). Added 8 C wrapper functions in
   wgpu_callbacks.c that accept callback info structs by pointer and
   dereference before forwarding to wgpu-native.

2. wgpuInstanceWaitAny unimplemented: Switched all sync operations from
   WaitAnyOnly + WaitAny to AllowSpontaneous + ProcessEvents/DevicePoll.

3. ASAP destruction: Mojo eagerly destroys Movable types after last
   reference. Added lifetime pins (_ = var^) for all GPU resources
   that outlive their last Mojo-visible reference.

4. Enum values: Fixed WGPUBufferBindingType off-by-one error.
   Storage=3, ReadOnlyStorage=4 (not 2/3 as originally coded).

5. Deprecated implicit conversion: Fixed UInt32 ternary in device.mojo.

Verified: compute_add.mojo runs end-to-end on NVIDIA RTX 3060,
all 1024 elements validated correct.

+++ wgpu/device.mojo
@@ -103,7 +103,7 @@ struct Device(Movable):
         label: String = "",
     ) raises -> Buffer:
         var label_sv = str_to_sv(label) if len(label) > 0 else WGPUStringView.null_view()
-        var mapped: UInt32 = 1 if mapped_at_creation else 0
+        var mapped: UInt32 = UInt32(1) if mapped_at_creation else UInt32(0)
         var desc_p = alloc[WGPUBufferDescriptor](1)
         desc_p[] = WGPUBufferDescriptor(
             OpaquePtr(),
